Get rack/parser working
Got the Rack::Parser code and custom parser working. Moved the branch
name parsing into its own method that is call from the call method.
Added a parse_deleted? method as well. The Sinatra params will now
contain params called branch(String) and deleted(Boolean). Code in the
payload POST route has been updated to use these params instead of
running the parsing logic itself.

I also created a new directory call "parsers" under "lib" and moved
webhook_json_parser.rb there. The class's namespace has been updated to
Sinatra::Parsers::WebhookJsonParser. All calls into that file/class have
been updated as well.

All "puts" in the code have been removed as they were there for Dev
troubleshooting purposes.

Updated .gitignore to ignore the .log files in the logs directory,
but also added a .keep file inside the logs directory so users/devs
don't have to create it themselves.

Gemfile.lock was updated after a bundle install and bin/puppet_webhook
had its permissions changed to 755.

All other work is related to the JSON parsing with Rack::Parser
